requirements.txt
Flask==2.3.3
mysql-connector-python
Flask-Login
Flask-WTF
WTForms
